Implemented hide items in menu by user type and by fragment name. Implemented hide buttons by user type.

+++ app/src/main/java/com/rydvi/clean_vrn/api/Session.kt
@@ -11,7 +11,7 @@ class Session {
         }
 
     @JsonProperty("id_user_type")
-    var idUserType: Int? = null
+    var idUserType: Long? = null
         get() = field
         set(value) {
             field = value
